backend/requirements.txt
fastapi==0.115.2
uvicorn==0.23.2
langchain==0.3.22
langchain-openai==0.3.12
langchain-core==0.3.51
openai==1.72.0
python-dotenv==1.0.0
pydantic==2.11.3